use an AsyncTaskLoader instead of AsyncTask in our GitHub query application
Found 20 TODO items in 1 file
com.example.android.asynctaskloader (20 items in 1 file)
MainActivity.java
(36, 4) // TODO (1) implement LoaderManager.LoaderCallbacks<String> on MainActivity
(43, 8) // TODO (28) Remove the key for storing the search results JSON
(47, 8) // TODO (2) Create a constant int to uniquely identify your loader. Call it GITHUB_SEARCH_LOADER
(80, 16) // TODO (26) Remove the code that retrieves the JSON
(84, 16) // TODO (25) Remove the code that displays the JSON
(88, 12) // TODO (24) Initialize the loader with GITHUB_SEARCH_LOADER as the ID, null for the bundle, and this for the context
(103, 12) // TODO (17) If no search was entered, indicate that there isn't anything to search for and return
(123, 12) // TODO (18) Remove the call to execute the AsyncTask
(127, 12) // TODO (19) Create a bundle called queryBundle
(130, 12) // TODO (20) Use putString with SEARCH_QUERY_URL_EXTRA as the key and the String value of the URL as the value
(133, 12) // TODO (21) Call getSupportLoaderManager and store it in a LoaderManager variable
(150, 12) // TODO (22) Get our Loader by calling getLoader and passing the ID we specified
(152, 12) // TODO (23) If the Loader was null, initialize it. Else, restart it.
(188, 8) // TODO (3) Override onCreateLoader
(243, 8) // TODO (13) Override onLoadFinished
(247, 12) // TODO (14) Hide the loading indicator
(250, 12) // TODO (15) Use the same logic used in onPostExecute to show the data or the error message
(264, 8) // TODO (16) Override onLoaderReset as it is part of the interface we implement, but don't do anything in this method
(274, 8) // TODO (29) Delete the AsyncTask class
(330, 12) // TODO (27) Remove the code that persists the JSON